+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a new helper method `DynamicMessage::decode`.
+
 ### Fixed
 
 - Fixed unpacked fields not accepting packed bytes